Refactoring API folder. Changed pathes to some api.
